rca_sky-main/requirements.txt
requests==2.31.0
pandas==2.0.3
numpy==1.24.3
# scikit-learn==1.3.0
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2
pyyaml==6.0.1
python-dateutil==2.8.2
schedule==1.2.0
ollama==0.2.1
aiohttp==3.8.5
networkx==3.1